@@ -16,21 +16,21 @@
   "private": true,
   "dependencies": {
     "@akveo/ng2-completer": "^9.0.1",
-    "@angular/animations": "^17.3.11",
-    "@angular/common": "^17.3.11",
-    "@angular/compiler": "^17.3.11",
-    "@angular/core": "^17.3.11",
-    "@angular/forms": "^17.3.11",
-    "@angular/localize": "^17.3.11",
-    "@angular/material": "^17.0.0",
-    "@angular/platform-browser": "^17.3.11",
-    "@angular/platform-browser-dynamic": "^17.3.11",
-    "@angular/router": "^17.3.11",
-    "@fortawesome/angular-fontawesome": "^0.14.0",
+    "@angular/animations": "^18.1.0",
+    "@angular/common": "^18.1.0",
+    "@angular/compiler": "^18.1.0",
+    "@angular/core": "^18.1.0",
+    "@angular/forms": "^18.1.0",
+    "@angular/localize": "^18.1.0",
+    "@angular/material": "^18.0.0",
+    "@angular/platform-browser": "^18.1.0",
+    "@angular/platform-browser-dynamic": "^18.1.0",
+    "@angular/router": "^18.1.0",
+    "@fortawesome/angular-fontawesome": "^0.15.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
-    "@ng-select/ng-select": "^12.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^17.0.0",
+    "@ng-select/ng-select": "^13.0.0",
     "@popperjs/core": "^2.11.8",
     "core-js": "^2.6.11",
     "jquery": "^3.4.1",
@@ -44,9 +44,9 @@
     "zone.js": "~0.14.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.8",
-    "@angular/cli": "^17.3.8",
-    "@angular/compiler-cli": "^17.3.11",
+    "@angular-devkit/build-angular": "^18.1.0",
+    "@angular/cli": "^18.1.0",
+    "@angular/compiler-cli": "^18.1.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/leaflet": "1.5.19",
