@@ -1,7 +1,7 @@
 import { BrowserModule } from "@angular/platform-browser";
 import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
-import { HttpClientModule } from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import { HashLocationStrategy, LocationStrategy } from "@angular/common";
 
 import { AppComponent } from "./app.component";
@@ -38,45 +38,39 @@ import { MatCheckboxModule } from "@angular/material/checkbox";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { BreadcrumbComponent } from "./breadcrumb/breadcrumb.component";
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    HeaderComponent,
-    FooterComponent,
-    ProjectComponent,
-    ProjectsComponent,
-    SearchComponent,
-    CscsComponent,
-    TopicsComponent,
-    MapComponent,
-    SearchNavComponent,
-    CscComponent,
-    ProjectResourceComponent,
-    BreadcrumbComponent,
-  ],
-  imports: [
-    BrowserModule,
-    FormsModule,
-    HttpClientModule,
-    routing,
-    NgSelectModule,
-    NgbDropdownModule,
-    NgbModule,
-    FontAwesomeModule,
-    MatTableModule,
-    MatSortModule,
-    MatCheckboxModule,
-    BrowserAnimationsModule,
-  ],
-  providers: [
-    GoogleAnalyticsService,
-    UmamiService,
-    LocalJsonService,
-    SciencebaseService,
-    SearchService,
-    UrlService,
-    { provide: LocationStrategy, useClass: HashLocationStrategy },
-  ],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [
+        AppComponent,
+        HeaderComponent,
+        FooterComponent,
+        ProjectComponent,
+        ProjectsComponent,
+        SearchComponent,
+        CscsComponent,
+        TopicsComponent,
+        MapComponent,
+        SearchNavComponent,
+        CscComponent,
+        ProjectResourceComponent,
+        BreadcrumbComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        FormsModule,
+        routing,
+        NgSelectModule,
+        NgbDropdownModule,
+        NgbModule,
+        FontAwesomeModule,
+        MatTableModule,
+        MatSortModule,
+        MatCheckboxModule,
+        BrowserAnimationsModule], providers: [
+        GoogleAnalyticsService,
+        UmamiService,
+        LocalJsonService,
+        SciencebaseService,
+        SearchService,
+        UrlService,
+        { provide: LocationStrategy, useClass: HashLocationStrategy },
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule {}
